+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 
+## 1.0.1
+
+- Refactor
+
 ## 1.0.0
 
 - Initial release
-- Re-enables Split Personality (randomize employee on revive)
+- Re-enables Split Personality (randomize employee/character on revive)
 - Re-enables Butter Fingers (randomize weapons on revive)
-- Configurable weights and enable toggles
-- Optional modifier pool debug logging
+- Configurable enable toggles and spawn weights
+- Optional mission modifier pool debug logging